Move: 모든 카테고리 조회 기능 Dto 패키지 이동
