+++ user-api/src/main/java/com/biengual/userapi/swagger/category/SwaggerCategory.java
@@ -1,9 +1,6 @@
 package com.biengual.userapi.swagger.category;
 
-import java.util.List;
-import java.util.Map;
-
-import com.biengual.userapi.category.domain.dto.CategoryResponseDto;
+import com.biengual.userapi.category.presentation.CategoryResponseDto;
 import com.biengual.userapi.swagger.SwaggerReturnInterface;
 
 public class SwaggerCategory extends SwaggerReturnInterface<CategoryResponseDto.AllCategoriesRes> {
